Change the alignment of the buffer to an u8 which allows the serialization of i8, i16 etc... to work.

+++ rust/interop_reserializer/julia_samples/types_tests.jl
@@ -5,11 +5,11 @@
 end
 
 mutable struct BookUpdate
-    time::Int64
-    timestamp_exch::Optional{Int64}
+    time::Int8
+    timestamp_exch::Optional{Int32}
     instId::Int64
-    updateType::Int64
+    updateType::Int128
 
-    bids::Vector{Int64}
+    bids::Vector{Int32}
     asks::Vector{Int64}
 end
